refactor:포스트 컨트롤러 부분에 ApiResponse와 ApiPageResponse로 Response의 일관성 구현

+++ src/main/java/com/sparta/newsfeed14thfriday/domain/post/service/PostService.java
@@ -12,6 +12,7 @@ import com.sparta.newsfeed14thfriday.domain.post.entity.Post;
 import com.sparta.newsfeed14thfriday.domain.post.repository.PostRepository;
 import com.sparta.newsfeed14thfriday.domain.user.entity.User;
 import com.sparta.newsfeed14thfriday.domain.user.repository.UserRepository;
+import com.sparta.newsfeed14thfriday.exception.AlreadyDeletedException;
 import com.sparta.newsfeed14thfriday.exception.DeletedUserIdException;
 import com.sparta.newsfeed14thfriday.exception.EmailNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -121,6 +122,10 @@ public class PostService {
 //            // delete
 //            postRepository.deleteByPostId(postId);
 //        }
+        if(post.getDeleted())
+        {
+            throw new AlreadyDeletedException("이미 삭제된 포스트 입니다");
+        }
         post.delete();
     }
